test(connect): simplify e2e setup
