+++ packages/connect/e2e/__fixtures__/index.ts
@@ -1,72 +1,72 @@
-import applyFlags from './applyFlags';
-import applySettings from './applySettings';
-import binanceSignTransaction from './binanceSignTransaction';
-import binanceGetPublicKey from './binanceGetPublicKey';
-import binanceGetAddress from './binanceGetAddress';
-import cardanoGetAddress from './cardanoGetAddress';
-import cardanoGetAddressDerivations from './cardanoGetAddressDerivations';
-import cardanoGetNativeScriptHash from './cardanoGetNativeScriptHash';
-import cardanoGetPublicKey from './cardanoGetPublicKey';
-import cardanoSignTransaction from './cardanoSignTransaction';
-import changeLanguage from './changeLanguage';
-import composeTransaction from './composeTransaction';
-import eosGetPublicKey from './eosGetPublicKey';
-import eosSignTransaction from './eosSignTransaction';
-import ethereumGetAddress from './ethereumGetAddress';
-import ethereumGetPublicKey from './ethereumGetPublicKey';
-import ethereumSignMessage from './ethereumSignMessage';
-import ethereumSignTransaction from './ethereumSignTransaction';
-import ethereumSignTransactionEip155 from './ethereumSignTransactionEip155';
-import ethereumSignTransactionEip1559 from './ethereumSignTransactionEip1559';
-import ethereumSignTypedData from './ethereumSignTypedData';
-import ethereumVerifyMessage from './ethereumVerifyMessage';
-import getAccountDescriptor from './getAccountDescriptor';
-import getAccountInfo from './getAccountInfo';
-import getAddress from './getAddress';
-import getAddressMultisig from './getAddressMultisig';
-import getAddressSegwit from './getAddressSegwit';
-import getFeatures from './getFeatures';
-import getFirmwareHash from './getFirmwareHash';
-import getOwnershipId from './getOwnershipId';
-import getOwnershipProof from './getOwnershipProof';
-import getPublicKey from './getPublicKey';
-import getPublicKeyBip48 from './getPublicKeyBip48';
-import nemGetAddress from './nemGetAddress';
-import nemSignTransactionMosaic from './nemSignTransactionMosaic';
-import nemSignTransactionMultisig from './nemSignTransactionMultisig';
-import nemSignTransactionOthers from './nemSignTransactionOthers';
-import nemSignTransactionTransfer from './nemSignTransactionTransfer';
-import rippleGetAddress from './rippleGetAddress';
-import rippleSignTransaction from './rippleSignTransaction';
-import signMessage from './signMessage';
-import signTransaction from './signTransaction';
-import signTransactionBcash from './signTransactionBcash';
-import signTransactionBech32 from './signTransactionBech32';
-import signTransactionBgold from './signTransactionBgold';
-import signTransactionDash from './signTransactionDash';
-import signTransactionDecred from './signTransactionDecred';
-import signTransactionDoge from './signTransactionDoge';
-import signTransactionExternal from './signTransactionExternal';
-import signTransactionKomodo from './signTransactionKomodo';
-import signTransactionMultisig from './signTransactionMultisig';
-import signTransactionMultisigChange from './signTransactionMultisigChange';
-import signTransactionPaymentRequest from './signTransactionPaymentRequest';
-import signTransactionPeercoin from './signTransactionPeercoin';
-import signTransactionReplace from './signTransactionReplace';
-import signTransactionSegwit from './signTransactionSegwit';
-import signTransactionTaproot from './signTransactionTaproot';
-import signTransactionZcash from './signTransactionZcash';
-import solanaGetAddress from './solanaGetAddress';
-import solanaGetPublicKey from './solanaGetPublicKey';
-import solanaSignTransaction from './solanaSignTransaction';
-import stellarGetAddress from './stellarGetAddress';
-import stellarSignTransaction from './stellarSignTransaction';
-import tezosGetAddress from './tezosGetAddress';
-import tezosGetPublicKey from './tezosGetPublicKey';
-import tezosSignTransaction from './tezosSignTransaction';
-import verifyMessage from './verifyMessage';
-import verifyMessageSegwit from './verifyMessageSegwit';
-import verifyMessageSegwitNative from './verifyMessageSegwitNative';
+export { default as applyFlags } from './applyFlags';
+export { default as applySettings } from './applySettings';
+export { default as binanceSignTransaction } from './binanceSignTransaction';
+export { default as binanceGetPublicKey } from './binanceGetPublicKey';
+export { default as binanceGetAddress } from './binanceGetAddress';
+export { default as cardanoGetAddress } from './cardanoGetAddress';
+export { default as cardanoGetAddressDerivations } from './cardanoGetAddressDerivations';
+export { default as cardanoGetNativeScriptHash } from './cardanoGetNativeScriptHash';
+export { default as cardanoGetPublicKey } from './cardanoGetPublicKey';
+export { default as cardanoSignTransaction } from './cardanoSignTransaction';
+export { default as changeLanguage } from './changeLanguage';
+export { default as composeTransaction } from './composeTransaction';
+export { default as eosGetPublicKey } from './eosGetPublicKey';
+export { default as eosSignTransaction } from './eosSignTransaction';
+export { default as ethereumGetAddress } from './ethereumGetAddress';
+export { default as ethereumGetPublicKey } from './ethereumGetPublicKey';
+export { default as ethereumSignMessage } from './ethereumSignMessage';
+export { default as ethereumSignTransaction } from './ethereumSignTransaction';
+export { default as ethereumSignTransactionEip155 } from './ethereumSignTransactionEip155';
+export { default as ethereumSignTransactionEip1559 } from './ethereumSignTransactionEip1559';
+export { default as ethereumSignTypedData } from './ethereumSignTypedData';
+export { default as ethereumVerifyMessage } from './ethereumVerifyMessage';
+export { default as getAccountDescriptor } from './getAccountDescriptor';
+export { default as getAccountInfo } from './getAccountInfo';
+export { default as getAddress } from './getAddress';
+export { default as getAddressMultisig } from './getAddressMultisig';
+export { default as getAddressSegwit } from './getAddressSegwit';
+export { default as getFeatures } from './getFeatures';
+export { default as getFirmwareHash } from './getFirmwareHash';
+export { default as getOwnershipId } from './getOwnershipId';
+export { default as getOwnershipProof } from './getOwnershipProof';
+export { default as getPublicKey } from './getPublicKey';
+export { default as getPublicKeyBip48 } from './getPublicKeyBip48';
+export { default as nemGetAddress } from './nemGetAddress';
+export { default as nemSignTransactionMosaic } from './nemSignTransactionMosaic';
+export { default as nemSignTransactionMultisig } from './nemSignTransactionMultisig';
+export { default as nemSignTransactionOthers } from './nemSignTransactionOthers';
+export { default as nemSignTransactionTransfer } from './nemSignTransactionTransfer';
+export { default as rippleGetAddress } from './rippleGetAddress';
+export { default as rippleSignTransaction } from './rippleSignTransaction';
+export { default as signMessage } from './signMessage';
+export { default as signTransaction } from './signTransaction';
+export { default as signTransactionBcash } from './signTransactionBcash';
+export { default as signTransactionBech32 } from './signTransactionBech32';
+export { default as signTransactionBgold } from './signTransactionBgold';
+export { default as signTransactionDash } from './signTransactionDash';
+export { default as signTransactionDecred } from './signTransactionDecred';
+export { default as signTransactionDoge } from './signTransactionDoge';
+export { default as signTransactionExternal } from './signTransactionExternal';
+export { default as signTransactionKomodo } from './signTransactionKomodo';
+export { default as signTransactionMultisig } from './signTransactionMultisig';
+export { default as signTransactionMultisigChange } from './signTransactionMultisigChange';
+export { default as signTransactionPaymentRequest } from './signTransactionPaymentRequest';
+export { default as signTransactionPeercoin } from './signTransactionPeercoin';
+export { default as signTransactionReplace } from './signTransactionReplace';
+export { default as signTransactionSegwit } from './signTransactionSegwit';
+export { default as signTransactionTaproot } from './signTransactionTaproot';
+export { default as signTransactionZcash } from './signTransactionZcash';
+export { default as solanaGetAddress } from './solanaGetAddress';
+export { default as solanaGetPublicKey } from './solanaGetPublicKey';
+export { default as solanaSignTransaction } from './solanaSignTransaction';
+export { default as stellarGetAddress } from './stellarGetAddress';
+export { default as stellarSignTransaction } from './stellarSignTransaction';
+export { default as tezosGetAddress } from './tezosGetAddress';
+export { default as tezosGetPublicKey } from './tezosGetPublicKey';
+export { default as tezosSignTransaction } from './tezosSignTransaction';
+export { default as verifyMessage } from './verifyMessage';
+export { default as verifyMessageSegwit } from './verifyMessageSegwit';
+export { default as verifyMessageSegwitNative } from './verifyMessageSegwitNative';
 
 // TODO: add fixtures for missing dependencies https://github.com/trezor/trezor-suite/issues/5353
 // backupDevice
@@ -94,96 +94,3 @@ import verifyMessageSegwitNative from './verifyMessageSegwitNative';
 // setProxy
 // tezosSignTransaction
 // wipeDevice
-
-let fixtures = [
-    applyFlags,
-    applySettings,
-    binanceGetAddress,
-    binanceGetPublicKey,
-    binanceSignTransaction,
-    cardanoGetAddress,
-    cardanoGetAddressDerivations,
-    cardanoGetNativeScriptHash,
-    cardanoGetPublicKey,
-    cardanoSignTransaction,
-    changeLanguage,
-    composeTransaction,
-    eosGetPublicKey,
-    eosSignTransaction,
-    ethereumGetAddress,
-    ethereumGetPublicKey,
-    ethereumSignMessage,
-    ethereumSignTransaction,
-    ethereumSignTransactionEip155,
-    ethereumSignTransactionEip1559,
-    ethereumSignTypedData,
-    ethereumVerifyMessage,
-    getAccountDescriptor,
-    getAccountInfo,
-    getAddress,
-    getAddressMultisig,
-    getAddressSegwit,
-    getFeatures,
-    getFirmwareHash,
-    getOwnershipId,
-    getOwnershipProof,
-    getPublicKey,
-    getPublicKeyBip48,
-    nemGetAddress,
-    nemSignTransactionMosaic,
-    nemSignTransactionMultisig,
-    nemSignTransactionOthers,
-    nemSignTransactionTransfer,
-    rippleGetAddress,
-    rippleSignTransaction,
-    signMessage,
-    signTransaction,
-    signTransactionBcash,
-    signTransactionBech32,
-    signTransactionBgold,
-    signTransactionDash,
-    signTransactionDecred,
-    signTransactionDoge,
-    signTransactionExternal,
-    signTransactionKomodo,
-    signTransactionMultisig,
-    signTransactionMultisigChange,
-    signTransactionPaymentRequest,
-    signTransactionPeercoin,
-    signTransactionReplace,
-    signTransactionSegwit,
-    signTransactionTaproot,
-    signTransactionZcash,
-    solanaGetAddress,
-    solanaGetPublicKey,
-    solanaSignTransaction,
-    stellarGetAddress,
-    stellarSignTransaction,
-    tezosGetAddress,
-    tezosGetPublicKey,
-    tezosSignTransaction,
-    verifyMessage,
-    verifyMessageSegwit,
-    verifyMessageSegwitNative,
-];
-
-const includedMethods = process.env.TESTS_INCLUDED_METHODS;
-const excludedMethods = process.env.TESTS_EXCLUDED_METHODS;
-if (includedMethods) {
-    const methodsArr = includedMethods.split(',');
-    fixtures = fixtures.filter(f => methodsArr.some(includedM => includedM === f.method));
-} else if (excludedMethods) {
-    const methodsArr = excludedMethods.split(',');
-    fixtures = fixtures.filter(f => !methodsArr.includes(f.method));
-}
-
-// sort by mnemonic to avoid emu re-loading
-const result = fixtures.sort((a, b) => {
-    if (!a.setup.mnemonic || !b.setup.mnemonic) return 0;
-    if (a.setup.mnemonic > b.setup.mnemonic) return 1;
-    if (b.setup.mnemonic > a.setup.mnemonic) return -1;
-
-    return 0;
-});
-
-export default result;
